Fix signal event error
When an error occurs, the MantisBT error handler clear all callbacks
(EVENT_GROUP_ACCESS_HAS_BUG_LEVEL) with `event_clear_callbacks()`.

When callbacks are cleared, `event_signal` returns the second parameter
that we have given to it. After that, `$p_user_id` becomes an array and
we don't want that so we put back the initial `$p_user_id` in it.

This patch fixes issue #3